backend/pptx_engine/requirements.txt
python-pptx>=1.0.2
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
lxml>=5.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
